Refactor client APIs
